feat!(controllers - db): Added RabbitMQ integration for tweet message queueing
this decision was made mostly to improve the loads on the server and have some kind of asynchronous response to the tweets, regardless of the number of tweets, because when a user sends a tweet, the Go app now can publish the tweet message to a RabbitMQ exchange.
The exchange can route the tweet messages to a queue, which can be consumed by a worker process.
The worker process can then extract the relevant information like (id, message, date) from the tweet message and store it in the MongoDB Atlas database.